normalization change
altered equation for normalization such that the lightcurve is centered around the median value, which is roughly 1. Also altered some hard-coded ratios for star in units of jupiter radii so we use the astropy.const package. Note that the overestimate of planet radius still occurs and is actually closer to 9 compared to the older value of 6